refactor skjema og gjør labels styrt av skjematype
